requirements.txt
langchain
langchain_openai
langgraph 
pydantic 
python-dotenv